% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_bootstrap_funcs.R
\name{bootstrapNull}
\alias{bootstrapNull}
\title{Bootstrap null distribution of F statistics for FDR estimation}
\usage{
bootstrapNull(df, maxit = 500, optim_fun_h0 = min_RSS_h0,
  optim_fun_h1 = min_RSS_h1, optim_fun_h1_2 = NULL, gr_fun_h0 = NULL,
  gr_fun_h1 = NULL, gr_fun_h1_2 = NULL, seed = NULL, ncores = 1,
  B = 3)
}
\arguments{
\item{df}{tidy data_frame retrieved after import of a 2D-TPP 
dataset, potential filtering and addition of a column "nObs"
containing the number of observations per protein}

\item{maxit}{maximal number of iterations the optimization
should be given, default is set to 500}

\item{optim_fun_h0}{optimization function that should be used
for fitting the H0 model}

\item{optim_fun_h1}{optimization function that should be used
for fitting the H1 model}

\item{optim_fun_h1_2}{optional additional optimization function 
that will be run with paramters retrieved from optim_fun_h1 and 
should be used for fitting the H1 model with the trimmed sum
model, default is NULL}

\item{gr_fun_h0}{optional gradient function for optim_fun_h0,
default is NULL}

\item{gr_fun_h1}{optional gradient function for optim_fun_h1,
default is NULL}

\item{gr_fun_h1_2}{optional gradient function for optim_fun_h1_2,
default is NULL}

\item{seed}{seed to set, default NULL corresponds to no seed set}

\item{ncores}{numeric value of numbers of cores that the function 
should use to parallelize}

\item{B}{numeric value of rounds of bootstrap, default: 3}
}
\value{
data frame containing F statistics of proteins with 
permuted 2D thermal profiles that are informative on the Null
distribution of F statistics
}
\description{
Bootstrap null distribution of F statistics for FDR estimation
}
\examples{
data("simulated_cell_extract_df")
temp_df <- simulated_cell_extract_df \%>\% 
  filter(clustername \%in\% paste0("protein", 1:20)) \%>\% 
  group_by(representative) \%>\% 
  mutate(nObs = n()) \%>\% 
  ungroup 
bootstrapNull(temp_df, B = 1)  

}
