% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_funcs.R
\name{computeFdr-defunct}
\alias{computeFdr-defunct}
\title{Compute FDR for given F statistics based on true and
null dataset (old function)}
\description{
Compute FDR for given F statistics based on true and
null dataset (old function)
}
\seealso{
\code{\link{TPP2D-defunct}}
}
\keyword{internal}
