% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_funcs.R
\name{computeFdr}
\alias{computeFdr}
\title{Compute FDR for given F statistics based on true and
null dataset (old function)}
\arguments{
\item{df_out}{data frame containing results from analysis by
fitAndEvalDataset}

\item{df_null}{data frame containing results from analysis by
bootstrapNull}
}
\value{
data frame annotating each protein with a FDR based on 
it's F statistic and number of observations
}
\description{
Compute FDR for given F statistics based on true and
null dataset (old function)
}
\seealso{
\code{\link{TPP2D-deprecated}}
}
\keyword{internal}