% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_analysis.R
\name{computeFStatFromParams}
\alias{computeFStatFromParams}
\title{Compute F statistics from paramter data frame}
\usage{
computeFStatFromParams(params_df, in_df = NULL,
  shrinkToPerTemperatureFstats = FALSE)
}
\arguments{
\item{params_df}{data frame listing all null and alternative
model parameters as obtained by 'getModelParamsDf'}

\item{in_df}{data frame of 2D-TPP profiles obtaine after 
data import}

\item{shrinkToPerTemperatureFstats}{logical indicating whether
F statistic should be shrunked towards per Temperature
F statistics}
}
\value{
data frame of all proteins and computed F statistics
and parameters that were used for the computation
}
\description{
Compute F statistics from paramter data frame
}
\examples{
data("simulated_cell_extract_df")
params_df <- getModelParamsDf(simulated_cell_extract_df)
computeFStatFromParams(params_df)
 
}
