% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_funcs.R
\name{computePvalFromKernelDensity}
\alias{computePvalFromKernelDensity}
\title{Compute p-values for given F statistics based on true and
null dataset using a kernel density estimation of the 
bootstrapped null distribution}
\usage{
computePvalFromKernelDensity(df_out, df_null)
}
\arguments{
\item{df_out}{data frame containing results from analysis by
fitAndEvalDataset}

\item{df_null}{data frame containing results from analysis by
bootstrapNull}
}
\value{
data frame annotating each protein with a FDR based on 
it's F statistic and number of observations
}
\description{
Compute p-values for given F statistics based on true and
null dataset using a kernel density estimation of the 
bootstrapped null distribution
}
\examples{
data("simulated_cell_extract_df")
temp_df <- simulated_cell_extract_df \%>\% 
  filter(clustername \%in\% paste0("protein", 1:3)) \%>\% 
  group_by(representative) \%>\% 
  mutate(nObs = n()) \%>\% 
  ungroup 
example_out <- fitAndEvalDataset(temp_df)
example_null <- bootstrapNull(temp_df, B = 2)
computePvalFromKernelDensity(
  example_out, 
  example_null)
 
}
