% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{config_tab}
\alias{config_tab}
\title{Example config table for a import of a simulated 
2D-TPP cell extract dataset}
\format{"Compound" describing the compound used for the assay, 
"Experiment" listing MS experiment ids of the separate runs
(typically comprising two multiplexed adjacent temperature), 
"Temperature": the temperature used for a given sub-experimet, 
the respective TMT labels "126"-"131L", RefCol referring to
the label used as a reference label for computing relative
fold changes (usually the label used for the control treatment).}
\description{
Config table fot import of simulated example 
dataset obtained by 2D-TPP experiments for analysis by 
the TPP2D-package. It's a data frame with the columns
"Compound" describing the compound used for the assay, 
"Experiment" listing MS experiment ids of the separate runs
(typically comprising two multiplexed adjacent temperature), 
"Temperature": the temperature used for a given sub-experimet, 
the respective TMT labels "126"-"131L", RefCol referring to
the label used as a reference label for computing relative
fold changes (usually the label used for the control treatment).
Please note that when the data is not supplied as a list of
already imported data frames the config table for the import 
function should be a path to an txt, csv or xlsx file containing
an additional column "Path" listing for each row the respective
path to a searched protein output file.
}
\examples{
data("config_tab")
data("raw_dat_list")
import_df <- import2dDataset(configTable = config_tab, 
data = raw_dat_list,
idVar = "protein_id",
intensityStr = "signal_sum_",
fcStr = "rel_fc_",
nonZeroCols = "qusm",
geneNameVar = "gene_name",
addCol = NULL,
qualColName = "qupm",
naStrs = c("NA", "n/d", "NaN"),
concFactor = 1e6,
medianNormalizeFC = TRUE,
filterContaminants = TRUE)
}
