% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_funcs.R
\name{configWide2Long}
\alias{configWide2Long}
\title{Tranform configuration table from wide to long}
\usage{
configWide2Long(configWide)
}
\arguments{
\item{configWide}{data frame containing a config table}
}
\value{
data frame containing config table in long format
}
\description{
Tranform configuration table from wide to long
}
\examples{
data("config_tab")
configWide2Long(configWide = config_tab)

}
