% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_funcs.R
\name{filterOutContaminants}
\alias{filterOutContaminants}
\title{Filter out contaminants}
\usage{
filterOutContaminants(dataLong)
}
\arguments{
\item{dataLong}{long format data frame of imported dataset}
}
\value{
data frame containing full dataset filtered to 
contain no contaminants
}
\description{
Filter out contaminants
}
\examples{
data("simulated_cell_extract_df")
filterOutContaminants(simulated_cell_extract_df)

}
