% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_analysis.R
\name{fitH1Model}
\alias{fitH1Model}
\title{Fit H1 model and evaluate fit statistics}
\usage{
fitH1Model(df, maxit = 500, optim_fun = min_RSS_h1,
  optim_fun_2 = NULL, gr_fun = NULL,
  ec50_lower_limit = min(unique(df$log_conc)[which(is.finite(unique(df$log_conc)))]),
  ec50_upper_limit = max(unique(df$log_conc)[which(is.finite(unique(df$log_conc)))]))
}
\arguments{
\item{df}{tidy data_frame retrieved after import of a 2D-TPP 
dataset, potential filtering and addition of a column "nObs"
containing the number of observations per protein}

\item{maxit}{maximal number of iterations the optimization
should be given, default is set to 500}

\item{optim_fun}{optimization function that should be used
for fitting the H0 model}

\item{optim_fun_2}{optional secound optimization function for
fitting the H1 model that should be used based on the fitted 
parameters of the optimizationfor based on optim_fun}

\item{gr_fun}{optional gradient funtion for optimization}

\item{ec50_lower_limit}{lower limit of ec50 parameter}

\item{ec50_upper_limit}{lower limit of ec50 parameter}
}
\value{
data frame with H1 model characteristics for each
protein
}
\description{
Fit H1 model and evaluate fit statistics
}
\examples{

data("simulated_cell_extract_df")
temp_df <- simulated_cell_extract_df \%>\% 
  filter(clustername \%in\% paste0("protein", 1:20)) \%>\% 
  group_by(representative) \%>\% 
  mutate(nObs = n()) \%>\% 
  ungroup
  
fitH1Model(temp_df)

fitH1Model(temp_df,
           optim_fun_2 = 
             TPP2D:::min_RSS_h1_trim)

}
