% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_filtering.R
\name{moderateOutliers}
\alias{moderateOutliers}
\title{Moderate outlier measurements}
\usage{
moderateOutliers(df, outlier_quantile = 0.98, qualColName = "qupm")
}
\arguments{
\item{df}{tidy data frame retrieved after import of a 2D-TPP 
dataset}

\item{outlier_quantile}{quantile that should be used as a cutoff to 
define outliers, default: 0.98}

\item{qualColName}{column indicating how many unique peptides 
were used for protein quantification}
}
\value{
data frame containing outlier moderated 2D-TPP data
}
\description{
Moderate outlier measurements
}
\examples{
data("simulated_cell_extract_df")
moderateOutliers(simulated_cell_extract_df) 
}
