% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot2dTppRelProfile}
\alias{plot2dTppRelProfile}
\title{Plot 2D thermal profile ratios of a protein}
\usage{
plot2dTppRelProfile(df, name)
}
\arguments{
\item{df}{tidy data frame of a 2D-TPP dataset}

\item{name}{gene name (clustername) of protein that 
should be visualized}
}
\value{
A ggplot displaying the thermal profile ratios of
a protein of choice in a datset of choice
}
\description{
Plot 2D thermal profile ratios of a protein
}
\examples{

data("simulated_cell_extract_df")
plot2dTppRelProfile(simulated_cell_extract_df, "protein1")

}
