% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{recomputeSignalFromRatios}
\alias{recomputeSignalFromRatios}
\title{Recompute robust signal intensities based on bootstrapped TMT channel ratios}
\usage{
recomputeSignalFromRatios(df)
}
\arguments{
\item{df}{tidy data_frame retrieved after import of a 2D-TPP dataset}
}
\value{
A data_frame with recomputed signal intensities (columname: value) 
and log2 transformed signal intensities (columnanme: log2_value) that more 
reliably reflect relative ratios between the TMT channels
}
\description{
Recompute robust signal intensities based on bootstrapped TMT channel ratios
}
\examples{

data("simulated_cell_extract_df")
recomputeSignalFromRatios(simulated_cell_extract_df)

}
