% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_cell_extract_df}
\alias{simulated_cell_extract_df}
\title{Example subset of a simulated 2D-TPP cell 
extract dataset}
\format{
data frame with columns representative (protein id), 
clustername (gene name), temperature, log_conc, raw_value, 
rel_value, value and log2_value
}
\usage{
data("simulated_cell_extract_df")
}
\description{
Simulated example dataset obtained by 2D-TPP 
experiments for analysis by the TPP2D-package. It contains 
a tidy data frame after import and recomputing of robust 
signal intensities with 200 simulated protein profiles 
(protein1-200) and 3 spiked-in true positives (TP1-3)
}
\keyword{datasets}
