% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPP2D-defunct.R, R/fdr_funcs.R
\name{TPP2D-defunct}
\alias{TPP2D-defunct}
\alias{computeFdr}
\title{Defunct functions in package \pkg{TPP2D}.}
\usage{
computeFdr(df_out, df_null)
}
\description{
The functions listed below are defunct and will be removed in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("-deprecated")}.
}
\details{
TPP2D defunct functions
}
\section{\code{computeFdr}}{

For \code{computeFdr}, use \code{\link{getFDR}}.
}

\keyword{internal}
