% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dExperiment-class.R
\docType{class}
\name{tpp2dExperiment-class}
\alias{tpp2dExperiment-class}
\alias{tpcaResult}
\title{S4 TPP2D Experiment Class}
\value{
an object of class tpp2dExperiment
}
\description{
S4 TPP2D Experiment Class
}
\section{Slots}{

\describe{
\item{\code{configTable}}{data.frame.}

\item{\code{idVar}}{character.}

\item{\code{intensityStr}}{character.}

\item{\code{fcStr}}{character.}

\item{\code{nonZeroCols}}{character.}

\item{\code{geneNameVar}}{character.}

\item{\code{qualColName}}{character.}

\item{\code{naStrs}}{character.}

\item{\code{concFactor}}{numeric.}

\item{\code{medianNormalizeFC}}{logical.}

\item{\code{filterContaminants}}{logical.}

\item{\code{minObs}}{numeric.}

\item{\code{independentFiltering}}{logical.}

\item{\code{fcThres}}{numeric.}

\item{\code{optim_fun_h0}}{function.}

\item{\code{optim_fun_h1}}{function.}

\item{\code{slopEC50}}{logical.}

\item{\code{maxit}}{numeric.}

\item{\code{BPPARAM}}{character.}

\item{\code{B}}{numeric}

\item{\code{byMsExp}}{logical.}

\item{\code{alpha}}{numeric.}

\item{\code{tidyDataTable}}{data.frame.}

\item{\code{modelParamsDf}}{data.frame}

\item{\code{resultTable}}{data.frame}

\item{\code{bootstrapNullDf}}{data.frame}

\item{\code{hitTable}}{data.frame}
}}

\examples{
tpp2dObj <- new("tpp2dExperiment")
}
