% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_funcs.R
\name{TPP_importCheckConfigTable}
\alias{TPP_importCheckConfigTable}
\title{Import and chech configuration table}
\usage{
TPP_importCheckConfigTable(infoTable, type = "2D")
}
\arguments{
\item{infoTable}{character string of a file path to
a config table (excel,txt or csv file) or data frame
containing a config table}

\item{type}{charater string indicating dataset type
default is 2D}
}
\value{
data frame with config table
}
\description{
Import and chech configuration table
}
\examples{
data("config_tab")
TPP_importCheckConfigTable(config_tab, type = "2D")
}
