% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_funcs.R
\name{computeFdr}
\alias{computeFdr}
\title{Compute FDR for given F statistics based on true and
null dataset}
\usage{
computeFdr(df_out, df_null)
}
\arguments{
\item{df_out}{data frame containing results from analysis by
fitAndEvalDataset}

\item{df_null}{data frame containing results from analysis by
bootstrapNull}
}
\value{
data frame annotating each protein with a FDR based on 
it's F statistic and number of obsevations
}
\description{
Compute FDR for given F statistics based on true and
null dataset
}
\examples{
data("simulated_cell_extract_df")
temp_df <- simulated_cell_extract_df \%>\% 
  filter(clustername \%in\% paste0("protein", 1:20)) \%>\% 
  group_by(representative) \%>\% 
  mutate(nObs = n()) \%>\% 
  ungroup 
example_out <- fitAndEvalDataset(temp_df)
example_null <- bootstrapNull(temp_df, B = 2)
computeFdr(example_out, example_null)
  
}
