% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_dat_list}
\alias{raw_dat_list}
\title{Example raw data for a subset of a simulated 
2D-TPP cell extract dataset}
\format{list of data frames with columns representative 
(protein id), clustername (gene name), temperature, log_conc, 
raw_value, rel_value, value and log2_value}
\usage{
raw_dat_list
}
\description{
Simulated example dataset obtained by 2D-TPP 
experiments for analysis by the TPP2D-package. It contains 
a list of data frames resembling raw data files returned 
from a MS database search with 200 simulated protein profiles 
(protein1-200) and 3 spiked-in true positives (TP1-3).
}
\examples{
data("raw_dat_list")
}
\keyword{datasets}
