% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpp2dExperiment-class.R
\docType{class}
\name{tpp2dExperiment-class}
\alias{tpp2dExperiment-class}
\alias{tpcaResult}
\title{S4 TPP2D Experiment Class}
\value{
an object of class tpp2dExperiment
with the following slots:
1) configFilePath:
}
\description{
S4 TPP2D Experiment Class
}
\section{Slots}{

\describe{
\item{\code{configFilePath}}{character.}

\item{\code{configTable}}{data.frame.}

\item{\code{tidyDataTable}}{data.frame.}

\item{\code{modelParamsDf}}{data.frame}

\item{\code{resultTable}}{}
}}

\examples{
tpp2dObj <- new("tpp2dExperiment", 
   configFilePath = system.file(package = "TPP2D"))
}
